third_party/blink/web_tests/virtual/synchronous_html_parser/media/video-play-pause-exception-expected.txt

CONSOLE ERROR: Uncaught (in promise) AbortError: The play() request was interrupted by a call to pause(). https://goo.gl/LdLk22
This is a testharness.js-based test.
PASS Test that the playing event is not fired when video has no src.
Harness: the test ran to completion.
